hmac: add test hmac NewWithEngine
